# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example Golang APIs",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/Expenses/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Expenses"
                ],
                "summary": "Create expenses",
                "parameters": [
                    {
                        "description": "Expense data",
                        "name": "req_dtos.Expenses",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_9cps_api-go-gin_dtos_request.Expenses"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.DefaultResponse"
                        }
                    }
                }
            }
        },
        "/HealthCheck/api": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HealthCheck"
                ],
                "summary": "Show status api",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.DefaultResponse"
                        }
                    }
                }
            }
        },
        "/HealthCheck/db": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HealthCheck"
                ],
                "summary": "Show status database",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.DefaultResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dtos.DefaultResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "date": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "github_com_9cps_api-go-gin_dtos_request.Expenses": {
            "type": "object",
            "properties": {
                "expensesMoney": {
                    "type": "number"
                },
                "expensesMonth": {
                    "type": "integer"
                },
                "expensesYear": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}